+++ src/hu_stuff.c
@@ -1296,7 +1296,10 @@ static void HU_widget_build_monsec(void)
 
   if ((hud_threelined_widgets && !st_crispyhud) || (st_crispyhud && nughud.sts_ml)) // [Nugget] NUGHUD
   {
-    sprintf(hud_monsecstr + offset, " \x1b%c%d%%", kills_percent_color, kills_percent);
+    // [Nugget]
+    if (hud_kills_percentage)
+    { sprintf(hud_monsecstr + offset, " \x1b%c%d%%", kills_percent_color, kills_percent); }
+
     HUlib_add_string_to_cur_line(&w_monsec, hud_monsecstr);
 
     sprintf(hud_monsecstr, "\x1b%cI \x1b%c%d/%d", ('0'+hudcolor_items), items_color, items, totalitems);
@@ -1307,9 +1310,12 @@ static void HU_widget_build_monsec(void)
   }
   else
   {
+    // [Nugget]
+    if (hud_kills_percentage)
+    { offset += sprintf(hud_monsecstr + offset, " \x1b%c%d%%", kills_percent_color, kills_percent); }
+
     sprintf(hud_monsecstr + offset,
-      " \x1b%c%d%% \x1b%cI \x1b%c%d/%d \x1b%cS \x1b%c%d/%d",
-      kills_percent_color, kills_percent,
+      " \x1b%cI \x1b%c%d/%d \x1b%cS \x1b%c%d/%d",
       '0'+hudcolor_items, items_color, items, totalitems,
       '0'+hudcolor_secrets, secrets_color, secrets, totalsecret);