+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Ammo and Health icons;
   - Toggle to apply patch offsets.
 - **Further Extended HUD color customization**
+- **_Show Kills Percentage [in Stats display]_ setting**
 - **_Direct Vertical Aiming_ for melee attacks**
 - **_Disable Melee Snapping_ setting**